Add check that integrals are unity
Of course, they aren't.  So that means something is wrong.  B(alpha)
seems to be off by factor of pi.  H(alpha) is more complicated - off by
factor of two at tau = 0, but increasingly bad for larger tau

Also, make a start on calculating the mean intensity also.  But still
not finished.